+++ src/cookbook.rs
@@ -49,27 +49,28 @@ cargo run --example app_demo  # or any other example
 34. [Chart: Styles: Example of setting default chart styles](#chart-styles-example-of-setting-default-chart-styles)
 35. [Chart: Chart data table](#chart-chart-data-table)
 36. [Chart: Chart data tools](#chart-chart-data-tools)
-37. [Chart: Gauge Chart](#chart-gauge-chart)
-38. [Chart: Chartsheet](#chart-chartsheet)
-39. [Grouped Rows: Create a grouped row outline](#grouped-rows-create-a-grouped-row-outline)
-40. [Grouped Columns: Create a grouped column outline](#grouped-columns-create-a-grouped-column-outline)
-41. [Textbox: Inserting Checkboxes in worksheets](#textbox-inserting-checkboxes-in-worksheets)
-42. [Textbox: Inserting Textboxes in worksheets](#textbox-inserting-textboxes-in-worksheets)
-43. [Textbox: Ignore Excel cell errors](#textbox-ignore-excel-cell-errors)
-44. [Sparklines: simple example](#sparklines-simple-example)
-45. [Sparklines: advanced example](#sparklines-advanced-example)
-46. [Traits: Extending generic `write()` to handle user data types](#traits-extending-generic-write-to-handle-user-data-types)
-47. [Macros: Adding macros to a workbook](#macros-adding-macros-to-a-workbook)
-48. [Defined names: using user defined variable names in worksheets](#defined-names-using-user-defined-variable-names-in-worksheets)
-49. [Cell Protection: Setting cell protection in a worksheet](#cell-protection-setting-cell-protection-in-a-worksheet)
-50. [Document Properties: Setting document metadata properties for a workbook](#document-properties-setting-document-metadata-properties-for-a-workbook)
-51. [Document Properties: Setting the Sensitivity Label](#document-properties-setting-the-sensitivity-label)
-52. [Internal links: Creating a Table of Contents](#internal-links-creating-a-table-of-contents)
-53. [Headers and Footers: Shows how to set headers and footers](#headers-and-footers-shows-how-to-set-headers-and-footers)
-54. [Hyperlinks: Add hyperlinks to a worksheet](#hyperlinks-add-hyperlinks-to-a-worksheet)
-55. [Freeze Panes: Example of setting freeze panes in worksheets](#freeze-panes-example-of-setting-freeze-panes-in-worksheets)
-56. [Dynamic array formulas: Examples of dynamic arrays and formulas](#dynamic-array-formulas-examples-of-dynamic-arrays-and-formulas)
-57. [Excel `LAMBDA()` function: Example of using the Excel 365 `LAMBDA()` function](#excel-lambda-function-example-of-using-the-excel-365-lambda-function)
+37. [Chart: Clustered categories](#chart-clustered-categories)
+38. [Chart: Gauge Chart](#chart-gauge-chart)
+39. [Chart: Chartsheet](#chart-chartsheet)
+40. [Grouped Rows: Create a grouped row outline](#grouped-rows-create-a-grouped-row-outline)
+41. [Grouped Columns: Create a grouped column outline](#grouped-columns-create-a-grouped-column-outline)
+42. [Textbox: Inserting Checkboxes in worksheets](#textbox-inserting-checkboxes-in-worksheets)
+43. [Textbox: Inserting Textboxes in worksheets](#textbox-inserting-textboxes-in-worksheets)
+44. [Textbox: Ignore Excel cell errors](#textbox-ignore-excel-cell-errors)
+45. [Sparklines: simple example](#sparklines-simple-example)
+46. [Sparklines: advanced example](#sparklines-advanced-example)
+47. [Traits: Extending generic `write()` to handle user data types](#traits-extending-generic-write-to-handle-user-data-types)
+48. [Macros: Adding macros to a workbook](#macros-adding-macros-to-a-workbook)
+49. [Defined names: using user defined variable names in worksheets](#defined-names-using-user-defined-variable-names-in-worksheets)
+50. [Cell Protection: Setting cell protection in a worksheet](#cell-protection-setting-cell-protection-in-a-worksheet)
+51. [Document Properties: Setting document metadata properties for a workbook](#document-properties-setting-document-metadata-properties-for-a-workbook)
+52. [Document Properties: Setting the Sensitivity Label](#document-properties-setting-the-sensitivity-label)
+53. [Internal links: Creating a Table of Contents](#internal-links-creating-a-table-of-contents)
+54. [Headers and Footers: Shows how to set headers and footers](#headers-and-footers-shows-how-to-set-headers-and-footers)
+55. [Hyperlinks: Add hyperlinks to a worksheet](#hyperlinks-add-hyperlinks-to-a-worksheet)
+56. [Freeze Panes: Example of setting freeze panes in worksheets](#freeze-panes-example-of-setting-freeze-panes-in-worksheets)
+57. [Dynamic array formulas: Examples of dynamic arrays and formulas](#dynamic-array-formulas-examples-of-dynamic-arrays-and-formulas)
+58. [Excel `LAMBDA()` function: Example of using the Excel 365 `LAMBDA()` function](#excel-lambda-function-example-of-using-the-excel-365-lambda-function)
 
 
 # Hello World: Simple getting started example
@@ -5179,6 +5180,102 @@ fn main() -> Result<(), XlsxError> {
     Ok(())
 }
 ```
+
+
+# Chart: Clustered categories
+
+Example of creating a clustered Excel chart where there are two levels of
+category on the X axis.
+
+The categories in clustered charts are 2D ranges, instead of the more normal
+1D ranges.
+
+**Image of the output file:**
+
+<img src="https://rustxlsxwriter.github.io/images/app_chart_clustered.png">
+
+
+**Code to generate the output file:**
+
+```rust
+// Sample code from examples/app_chart_clustered.rs
+
+use rust_xlsxwriter::{Chart, ChartType, Format, Workbook, XlsxError};
+
+fn main() -> Result<(), XlsxError> {
+    let mut workbook = Workbook::new();
+    let worksheet = workbook.add_worksheet();
+    let bold = Format::new().set_bold();
+
+    // Add the worksheet data that the charts will refer to.
+    worksheet.write_with_format(0, 0, "Types", &bold)?;
+    worksheet.write_with_format(0, 1, "Sub Type", &bold)?;
+    worksheet.write_with_format(0, 2, "Value 1", &bold)?;
+    worksheet.write_with_format(0, 3, "Value 2", &bold)?;
+    worksheet.write_with_format(0, 4, "Value 3", &bold)?;
+
+    worksheet.write(1, 0, "Type 1")?;
+    worksheet.write(1, 1, "Sub Type A")?;
+    worksheet.write(2, 1, "Sub Type B")?;
+    worksheet.write(3, 1, "Sub Type C")?;
+
+    worksheet.write(4, 0, "Type 2")?;
+    worksheet.write(4, 1, "Sub Type D")?;
+    worksheet.write(5, 1, "Sub Type E")?;
+
+    worksheet.write(1, 2, 5000)?;
+    worksheet.write(2, 2, 2000)?;
+    worksheet.write(3, 2, 250)?;
+    worksheet.write(4, 2, 6000)?;
+    worksheet.write(5, 2, 500)?;
+
+    worksheet.write(1, 3, 8000)?;
+    worksheet.write(2, 3, 3000)?;
+    worksheet.write(3, 3, 1000)?;
+    worksheet.write(4, 4, 6500)?;
+    worksheet.write(5, 3, 300)?;
+
+    worksheet.write(1, 4, 6000)?;
+    worksheet.write(2, 4, 4000)?;
+    worksheet.write(3, 4, 2000)?;
+    worksheet.write(4, 3, 6000)?;
+    worksheet.write(5, 4, 200)?;
+
+    // Create a new chart object.
+    let mut chart = Chart::new(ChartType::Column);
+
+    // Configure the series. Note, that the categories are 2D ranges (from
+    // column A to column B). This creates the clusters.
+    chart
+        .add_series()
+        .set_categories("Sheet1!$A$2:$B$6")
+        .set_values("Sheet1!$C$2:$C$6");
+
+    chart
+        .add_series()
+        .set_categories("Sheet1!$A$2:$B$6")
+        .set_values("Sheet1!$D$2:$D$6");
+
+    chart
+        .add_series()
+        .set_categories("Sheet1!$A$2:$B$6")
+        .set_values("Sheet1!$E$2:$E$6");
+
+    // Set the Excel chart style.
+    chart.set_style(37);
+
+    // Turn off the legend.
+    chart.legend().set_hidden();
+
+    // Insert the chart into the worksheet.
+    worksheet.insert_chart(2, 6, &chart)?;
+
+    workbook.save("chart_clustered.xlsx")?;
+
+    Ok(())
+}
+```
+
 
 
 # Chart: Gauge Chart